Pipes::isCollision() now works.
We split the pipes into two -- top and bottom. The reason behind this is
we use the sfml function GetGlobalBounds() to check if the bird has
collided with the pipe. If we only use one pipe sprite instead of using a
top and bottom, the gap in the middle of the pipes would be considered
part of the border.

tl;dr: We use GetGlobalBounds() on the pipes to check if the bird is
within the bounds. Pipes::isCollision() works properly now.