Added game sounds and game states.
There are now three game states: preGame, midGame, and postGame.
During the preGame state, the bird simply flaps its wings and some music
plays. The midGame state is basically what we've been working on all
quarter, and the postGame state is just a screen that says "Game Over"
with nothing but the moving buildings in the background. In the preGame
state, if you press space, the bird jumps and the game begins, in the
postGame state, if you press space or enter, the game gets into the
midGame state and all the game's elements reset and you can play again
normaly.

I also added sound effects for scoring points and for the bird dying
(which I plan to tweak a little bit). Other than that, I added some music
that I thought would be hilariously annoying and a 50% opaque png that
gets displayed when the game is over.